+++ CHANGELOG.md
@@ -139,4 +139,4 @@
 * Added Reset functionality to SkippableTimerScript
 * Added option for ApplyThemeScript to apply when theme policy is ExplicitOnly or not
 * Cleaned up handling of base frame in dialogue scene (now available via DialogueScene.BaseFrame)
-
+* Suppressed warnings from ExplicitKBMInputMapper when noclipping (and other "missing axis" warnings as well)